car_management/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.8
filelock==3.16.1
idna==3.10
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
pillow==11.0.0
pipreqs==0.4.13
platformdirs==4.3.6
PyJWT==2.9.0
PyMySQL==1.1.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.26.6
yarg==0.1.10
